requirements.txt
opencv-python
numpy
tqdm
scikit-image
torch
torchvision
pandas
matplotlib
Flask
Pillow
